+++ pandas/core/arrays/sparse/__init__.py
@@ -8,7 +8,6 @@ from pandas.core.arrays.sparse.array import (
     SparseArray,
     make_sparse_index,
 )
-from pandas.core.arrays.sparse.dtype import SparseDtype
 
 __all__ = [
     "BlockIndex",
@@ -16,6 +15,5 @@ __all__ = [
     "make_sparse_index",
     "SparseAccessor",
     "SparseArray",
-    "SparseDtype",
     "SparseFrameAccessor",
 ]
