+++ pandas/tests/arrays/sparse/test_dtype.py
@@ -4,7 +4,7 @@ import numpy as np
 import pytest
 
 import pandas as pd
-from pandas.core.arrays.sparse import SparseDtype
+from pandas import SparseDtype
 
 
 @pytest.mark.parametrize(
